+++ excalidraw-app/App.tsx
@@ -871,7 +871,7 @@ const ExcalidrawWrapper = () => {
           <OverwriteConfirmDialog.Actions.SaveToDisk />
           {/* AstraDraw+: Disabled until user functionality is implemented */}
         </OverwriteConfirmDialog>
-        <AppFooter onChange={() => excalidrawAPI?.refresh()} />
+        <AppFooter onChange={() => excalidrawAPI?.refresh()} excalidrawAPI={excalidrawAPI} />
         {excalidrawAPI && <AIComponents excalidrawAPI={excalidrawAPI} />}
 
         <TTDDialogTrigger />